Added more scatters to moons
I'm getting this odd issue where im seeing massive numbers of floating
scatters. I am not really sure what the common factor is for the bug
occuring - i first noticed on Gloth that if i deleted the cache and
loaded the game, the scatters  would spawn correctly, but if i closed
and reopened the game without clearing the cache, i'd get a second sset
of scatters that were either floating or underground.  It seemed as if
they were following the heightmap, but it had been displaced a bit? By
this i mean that they seemed to be conforming to invisible landscape features
rather than just being placed at random heights. I
also observed this on Lema.  However, while the issue was also present
on Chople, Ollagander and Basil, it presented a little differently in that the
floating scatters formed these hundred-meter tall ribbons that snaked
across the landscape, not following any features anywhere on the
heightmap. I am however not quite ready to rule out that this is just me
making some mistake so i'll look at some other mods to see how they do
scatters.

Next step is momuments!