Changed Run into a contact binary
I was inspired by the Lucy mission to try making a contact binary type
object in Kopernicus.  Run was the obvious body to modify because, to be
blunt, it was ugly and boring, and i think its worked reasonably well.
The oblate PQS mod really helps with achieving the desired extreme
topography without too-too much visible banding from the heightmap,
which otherwise would totally wreck the entire look of the thing (you
can see this quite clearly in other popular planet packs, cough cough
Outer Planets).

The thing hitch is that the vertexHeight type mods don't play very well
with the oblate mod, which results in some unpleasant stretched
topographical features.  THis was a problem with the original version of
Run, and is maybe a BIT better now, but its still noticable on the
inner sides of the two bodies.  The only i can think to avoid this would
be to not use those mods at all and reply entirely on the heightmap for
surface detail, but that would be bad because small heightmap details
end up looking like pixelated messes.  So in the end i just gotta cope
and mald with it.

Look at me, waffling to no one.  Lol.  I will release the mod soon.