Actually got rid of the banding on Run
What i've discovered is that .dds files, which seem to be favored by
other planet pack creates, sort of just suck?  They tend to form really
ugly compression artifacts, like lots of little dots on areas of flat
colour and random discolourations.  And, as it turns out, banding.
Saving the heightmap as a png almost entirely cured the issue.  So there
ya go.  Live and learn.